requirements.txt
django=1.3.1
south=0.7.3
sorl-thumbnail